feat: IPS - first approach to the algorithm
Selection is based on the "beauty" contest of the nodes - each node is evaluated based on its degree,
betweenness, closeness and eigenvector centrality. Then, if requested ranking is updated with location factor.
Each factor has its own weight that is used to determine factor's importance to the calculation of the final ranking.

Weights are defined in the configuration file. When the ranking is calculated, the peers are selected based on the ranking.
The number of peers can be changed and it is defined in the configuration file.
Algorithm is constructed as step by step process, so it is easy to add new steps or change the order of the steps.

Currently params taken into account:
- node params (degree, betweenness, eigenvector, closeness)
- node location
- peerlist size
- hot and cold nodes elimination - partly, trying to avoid creating hot nodes by selecting nodes with lower betweenness

What is not taken into account (yet):
- performance of the node (we have no data for that)
- performance of the node's connection to other nodes (we've only connection time measured)
- island detection (tbd)